Break tests for overriding actions in the policy out into its own test file
