+++ test/base_policy_test.rb
@@ -45,63 +45,6 @@ describe Accessly::Policy::Base do
     # TODO: Revoke
   end
 
-  class CustomizedPolicy < UserPolicy
-
-    def self.namespace
-      "OverriddenNamespace"
-    end
-
-    def is_admin?
-      actor.admin?
-    end
-
-    # TODO: List
-    # def self.admin_scope
-    #   actor.company.users.select(:id)
-    # end
-
-    # def list
-    #   if actor.admin?
-    #     actor.company.users
-    #   else
-    #     super
-    #   end
-    # end
-
-    # Customize a general action check
-    def destroy?
-      if actor.name == "Aaron"
-        true
-      else
-        super
-      end
-    end
-
-    # Customize a check that is both general and on an object
-    def change_role?(object = nil)
-      if object.nil?
-        if actor.name == "Bob"
-          false
-        else
-          super
-        end
-      elsif actor.name == "Bob" && object.name == "Aaron"
-        true
-      else
-        super
-      end
-    end
-
-    # Customize an object action check
-    def email?(object)
-      if object.name == "Aaron"
-        true
-      else
-        super
-      end
-    end
-  end
-
   class DefaultNamespacePolicy < Accessly::Policy::Base
     actions view: 1
   end
@@ -210,60 +153,6 @@ describe Accessly::Policy::Base do
     policy.email?(other_user).must_equal true
     permission.destroy!
     policy.email?(other_user).must_equal true
-  end
-
-  it "allows general action checks to be customized" do
-    # User named Aaron can always destroy users
-    user = User.create!(name: "Aaron")
-    policy = CustomizedPolicy.new(user)
-    policy.destroy?.must_equal true
-
-    # User not named Aaron gets normal privileges
-    user = User.create!(name: "Jim")
-    policy = CustomizedPolicy.new(user)
-    policy.destroy?.must_equal false
-  end
-
-  it "allows object action checks to be customized" do
-    # Anybody can email user named Aaron
-    user = User.create!
-    other_user = User.create!(name: "Aaron")
-    policy = CustomizedPolicy.new(user)
-    policy.email?(other_user).must_equal true
-
-    # Emailing other users goes through normal privilege check
-    policy.email?(user).must_equal false
-  end
-
-  it "allows checks that are both general and on an object to be customized" do
-    # User named Bob cannot generally change role
-    user = User.create!(name: "Bob")
-    other_user = User.create!(name: "Aaron")
-    policy = CustomizedPolicy.new(user)
-    policy.change_role?.must_equal false
-
-    # User named Bob can change role for specific user named Aaron
-    policy.change_role?(other_user).must_equal true
-
-    # User named Aaron has normal privileges
-    policy = CustomizedPolicy.new(other_user)
-    policy.change_role?.must_equal false
-    policy.change_role?(user).must_equal false
-  end
-
-  it "returns true automatically when is_admin? returns true" do
-    admin_user = User.create!(admin: true)
-    non_admin_user = User.create!
-
-    # Non-admin has no permissions set
-    non_admin_policy = CustomizedPolicy.new(non_admin_user)
-    non_admin_policy.view?.must_equal false
-    non_admin_policy.view?(admin_user).must_equal false
-
-    # Admin has no permissions set, but can do anything
-    admin_policy = CustomizedPolicy.new(admin_user)
-    admin_policy.view?.must_equal true
-    admin_policy.view?(non_admin_user).must_equal true
   end
 
   it "uses the policy class name as the default namespace" do
